+++ sdk/builder.go
@@ -5,6 +5,8 @@ import (
 	"crypto/sha256"
 	"encoding/hex"
 	"fmt"
+	"math/rand"
+	"time"
 
 	"github.com/Masterminds/semver/v3"
 	"github.com/giantswarm/microerror"
@@ -16,16 +18,22 @@ const (
 	preReleaseHashLength = 10
 )
 
+// Builder builds custom Releases by overriding a base release with custom cluster app and default apps.
 type Builder struct {
 	client      *Client
 	provider    Provider
 	baseRelease string
 
-	preReleasePrefix string
-	clusterApp       *ReleaseSpecComponent
-	apps             []ReleaseSpecApp
+	// pre-release config
+	preRelease preReleaseConfig
+
+	// overrides
+	clusterApp *ReleaseSpecComponent
+	apps       []ReleaseSpecApp
 }
 
+// NewBuilder creates a new Builder with specified releases Client, provider and base release.
+// If a base release is an empty string it will use the latest available release for the provider.
 func NewBuilder(client *Client, provider Provider, baseRelease string) (*Builder, error) {
 	if client == nil {
 		return nil, microerror.Maskf(InvalidConfigError, "client must not be empty")
@@ -34,13 +42,18 @@ func NewBuilder(client *Client, provider Provider, baseRelease string) (*Builder
 		return nil, microerror.Maskf(UnsupportedProviderError, "provider `%s` is not supported", provider)
 	}
 
+	var overridesDiff diff
 	return &Builder{
 		client:      client,
 		provider:    provider,
 		baseRelease: baseRelease,
+		preRelease: preReleaseConfig{
+			overridesDiff: &overridesDiff,
+		},
 	}, nil
 }
 
+// WithClusterApp overrides the cluster-<provider> app in the base release.
 func (b *Builder) WithClusterApp(version, catalog string) *Builder {
 	b.clusterApp = &ReleaseSpecComponent{
 		Name:    fmt.Sprintf("cluster-%s", b.provider),
@@ -50,6 +63,7 @@ func (b *Builder) WithClusterApp(version, catalog string) *Builder {
 	return b
 }
 
+// WithApp overrides the default app in the base release.
 func (b *Builder) WithApp(name, version, catalog string, dependsOn []string) *Builder {
 	app := ReleaseSpecApp{
 		Name:      name,
@@ -61,48 +75,22 @@ func (b *Builder) WithApp(name, version, catalog string, dependsOn []string) *Bu
 	return b
 }
 
-func (b *Builder) WithPreReleasePrefix(prePrefix string) *Builder {
-	b.preReleasePrefix = prePrefix
+// WithPreReleasePrefix sets a custom prefix that is prepended to the pre-release of the custom release version.
+func (b *Builder) WithPreReleasePrefix(prefix string) *Builder {
+	b.preRelease.prefix = prefix
 	return b
 }
 
+// WithRandomPreRelease specifies that custom release version will have a random pre-release with specified length. The
+// specified length does not include the length of the optionally specified pre-release prefix.
+func (b *Builder) WithRandomPreRelease(length int) *Builder {
+	b.preRelease.isRandom = true
+	b.preRelease.length = length
+	return b
+}
+
+// Build a custom release.
 func (b *Builder) Build(ctx context.Context) (*Release, error) {
-	var buildPreRelease bool
-	if b.clusterApp != nil {
-		// Check if cluster app version is pre-release
-		clusterAppVersion, err := semver.NewVersion(b.clusterApp.Version)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-		buildPreRelease = clusterAppVersion.Prerelease() != ""
-	}
-	for _, app := range b.apps {
-		// Check if app version is pre-release
-		appVersion, err := semver.NewVersion(app.Version)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-		buildPreRelease = appVersion.Prerelease() != ""
-	}
-
-	release, err := b.build(ctx, buildPreRelease)
-	if err != nil {
-		return nil, microerror.Mask(err)
-	}
-
-	return release, nil
-}
-
-func (b *Builder) BuildPreRelease(ctx context.Context) (*Release, error) {
-	release, err := b.build(ctx, true)
-	if err != nil {
-		return nil, microerror.Mask(err)
-	}
-
-	return release, nil
-}
-
-func (b *Builder) build(ctx context.Context, buildPreRelease bool) (*Release, error) {
 	var release *Release
 	var err error
 
@@ -119,9 +107,12 @@ func (b *Builder) build(ctx context.Context, buildPreRelease bool) (*Release, er
 			return nil, microerror.Mask(err)
 		}
 	}
+	b.baseRelease, err = release.GetVersion()
+	if err != nil {
+		return nil, microerror.Mask(err)
+	}
 
 	versionBump := none
-	changes := ""
 
 	// Then we override cluster app
 	if b.clusterApp != nil {
@@ -137,7 +128,7 @@ func (b *Builder) build(ctx context.Context, buildPreRelease bool) (*Release, er
 			versionBump = clusterAppVersionBump
 		}
 		overrideClusterApp(release, *b.clusterApp)
-		changes += fmt.Sprintf("%s-%s-%s", fmt.Sprintf("cluster-%s", b.provider), b.clusterApp.Catalog, b.clusterApp.Version)
+		b.preRelease.overridesDiff.addComponentDiff(*b.clusterApp)
 	}
 
 	// And we override apps
@@ -154,10 +145,7 @@ func (b *Builder) build(ctx context.Context, buildPreRelease bool) (*Release, er
 			versionBump = appVersionBump
 		}
 		overrideApp(release, appOverride)
-		changes += fmt.Sprintf(";%s-%s-%s", appOverride.Name, appOverride.Catalog, appOverride.Version)
-		for _, appDependency := range appOverride.DependsOn {
-			changes += fmt.Sprintf("-%s", appDependency)
-		}
+		b.preRelease.overridesDiff.addAppDiff(appOverride)
 	}
 
 	// Now we get the base release version
@@ -188,11 +176,12 @@ func (b *Builder) build(ctx context.Context, buildPreRelease bool) (*Release, er
 		*releaseVersion = releaseVersion.IncPatch()
 	}
 
-	if buildPreRelease {
-		preRelease := hashAndTruncate(changes, preReleaseHashLength)
-		if b.preReleasePrefix != "" {
-			preRelease = fmt.Sprintf("%s.%s", b.preReleasePrefix, preRelease)
-		}
+	// Set pre-release is one is needed
+	preRelease, err := b.buildPreReleaseString()
+	if err != nil {
+		return nil, microerror.Mask(err)
+	}
+	if preRelease != "" {
 		*releaseVersion, err = releaseVersion.SetPrerelease(preRelease)
 		if err != nil {
 			return nil, microerror.Mask(err)
@@ -202,6 +191,83 @@ func (b *Builder) build(ctx context.Context, buildPreRelease bool) (*Release, er
 	// Finally, we update Release resource name
 	release.Name = fmt.Sprintf("%s-%s", b.provider, releaseVersion.String())
 	return release, nil
+}
+
+func (b *Builder) buildPreReleaseString() (string, error) {
+	if b.preRelease.length == 0 {
+		b.preRelease.length = 10
+	}
+	if b.preRelease.isRandom {
+		r := rand.New(rand.NewSource(time.Now().UnixNano()))
+		const charsWithoutZero = "abcdefghijklmnopqrstuvwxyz123456789"
+		const chars = "abcdefghijklmnopqrstuvwxyz0123456789"
+		result := make([]byte, b.preRelease.length)
+		// set first char to non-zero char
+		result[0] = charsWithoutZero[r.Intn(len(charsWithoutZero))]
+		// set remaining chars
+		for i := 1; i < b.preRelease.length; i++ {
+			result[i] = chars[r.Intn(len(chars))]
+		}
+		preReleaseString := string(result)
+		if b.preRelease.prefix != "" {
+			preReleaseString = fmt.Sprintf("%s.%s", b.preRelease.prefix, preReleaseString)
+		}
+		return preReleaseString, nil
+	}
+
+	// Since random pre-release is not needed, we now check if we need a pre-release or not.
+	// A pre-release string is needed if one of the following conditions is met:
+	// 1. Base release has a pre-release string, i.e. when you customise a base Release with a pre-release version
+	//    you get a custom Release with a pre-release version.
+	// 2. Cluster app override has a pre-release string, i.e. when you override a cluster app with a pre-release
+	//    version you get a custom Release with a pre-release version.
+	// 3. One of app overrides has a pre-release string, i.e. when you override an app with a pre-release version
+	//    you get a custom Release with pre-release version.
+
+	preReleaseBuildFunc := func() string {
+		preReleaseString := b.preRelease.overridesDiff.hashAndTruncate(b.preRelease.length)
+		if b.preRelease.prefix != "" {
+			preReleaseString = fmt.Sprintf("%s.%s", b.preRelease.prefix, preReleaseString)
+		}
+		return preReleaseString
+	}
+
+	// check base release
+	baseReleaseHasPreReleaseString, err := versionStringHasPreReleaseString(b.baseRelease)
+	if err != nil {
+		return "", microerror.Mask(err)
+	}
+	if baseReleaseHasPreReleaseString {
+		return preReleaseBuildFunc(), nil
+	}
+
+	// check if cluster app has a pre-release
+	clusterAppPreReleaseString, err := versionStringHasPreReleaseString(b.clusterApp.Version)
+	if err != nil {
+		return "", microerror.Mask(err)
+	}
+	if clusterAppPreReleaseString {
+		return preReleaseBuildFunc(), nil
+	}
+
+	// check if apps have a pre-release
+	for _, app := range b.apps {
+		appPreReleaseString, err := versionStringHasPreReleaseString(app.Version)
+		if err != nil {
+			return "", microerror.Mask(err)
+		}
+		if appPreReleaseString {
+			return preReleaseBuildFunc(), nil
+		}
+	}
+
+	// Finally, we return only pre-release prefix it has been specified
+	// check if pre-release prefix is specified
+	if b.preRelease.prefix != "" {
+		return b.preRelease.prefix, nil
+	}
+
+	return "", nil
 }
 
 func overrideClusterApp(release *Release, clusterApp ReleaseSpecComponent) {
@@ -230,23 +296,6 @@ func overrideApp(release *Release, appOverride ReleaseSpecApp) {
 		release.Spec.Apps[i] = appOverride
 		break
 	}
-}
-
-// hashAndTruncate takes a string that will be hashed and the desired length to which the hash result should be
-// truncated.
-//
-// It hashes the string using SHA-256 algorithm and returns the truncated hexadecimal representation
-// of the hash. If the length input is greater than or equal to the length of the hash, the entire
-// hash is returned.
-func hashAndTruncate(s string, length int) string {
-	h := sha256.New()
-	h.Write([]byte(s))
-	hashedString := hex.EncodeToString(h.Sum(nil))
-	if len(hashedString) < length {
-		length = len(hashedString)
-	}
-
-	return hashedString[:length]
 }
 
 type versionPart int
@@ -282,4 +331,61 @@ func getVersionPartToBump(newVersion, oldVersion string) (versionPart, error) {
 	}
 
 	return none, nil
+}
+
+// preReleaseConfig specifies how the pre-release is built.
+type preReleaseConfig struct {
+	// prefix of the pre-release string. e.g. v1.2.3-prefix.h97o23f46t
+	prefix string
+
+	// overridesDiff contains diff between apps and components from the base release and the apps and components that
+	// are overriding them.
+	overridesDiff *diff
+
+	// isRandom is a flag that indicates if the pre-release suffix is random.
+	isRandom bool
+
+	// length of the pre-release string (without prefix).
+	length int
+}
+
+type diff string
+
+func (d *diff) addAppDiff(app ReleaseSpecApp) {
+	newDiff := fmt.Sprintf("%s+%s+%s+%s", string(*d), app.Name, app.Catalog, app.Version)
+	for _, dependsOn := range app.DependsOn {
+		newDiff = fmt.Sprintf("%s+%s", newDiff, dependsOn)
+	}
+	*d = diff(newDiff)
+}
+
+func (d *diff) addComponentDiff(component ReleaseSpecComponent) {
+	newDiff := fmt.Sprintf("%s+%s+%s+%s", string(*d), component.Name, component.Catalog, component.Version)
+	*d = diff(newDiff)
+}
+
+// hashAndTruncate takes a string that will be hashed and the desired length to which the hash result should be
+// truncated.
+//
+// It hashes the string using SHA-256 algorithm and returns the truncated hexadecimal representation
+// of the hash. If the length input is greater than or equal to the length of the hash, the entire
+// hash is returned.
+func (d *diff) hashAndTruncate(length int) string {
+	h := sha256.New()
+	h.Write([]byte(*d))
+	hashedString := hex.EncodeToString(h.Sum(nil))
+	if len(hashedString) < length {
+		length = len(hashedString)
+	}
+
+	return hashedString[:length]
+}
+
+func versionStringHasPreReleaseString(v string) (bool, error) {
+	semverVersion, err := semver.NewVersion(v)
+	if err != nil {
+		return false, microerror.Mask(err)
+	}
+
+	return semverVersion.Prerelease() != "", nil
 }